+++ src/main/java/io/parallec/core/actor/PingWorker.java
@@ -298,7 +298,7 @@ public class PingWorker extends UntypedActor {
             final ResponseOnSingeRequest res = new ResponseOnSingeRequest(
                     response, failObtainResponse, errorMessage, stackTrace,
                     statusCode, statusCodeInt,
-                    PcDateUtils.getNowDateTimeStrStandard());
+                    PcDateUtils.getNowDateTimeStrStandard(), null);
 
             if (!getContext().system().deadLetters().equals(sender)) {
                 sender.tell(res, getSelf());