+++ src/test/java/io/parallec/core/pojo/ParallecPojoClassTest.java
@@ -28,6 +28,7 @@ import io.parallec.core.actor.message.ResponseFromManager;
 import io.parallec.core.actor.poll.HttpPollerProcessor;
 import io.parallec.core.actor.poll.PollerData;
 import io.parallec.core.bean.HttpMeta;
+import io.parallec.core.bean.ResponseHeaderMeta;
 import io.parallec.core.bean.SetAndCount;
 import io.parallec.core.bean.StrStrMap;
 import io.parallec.core.bean.TargetHostMeta;
@@ -160,7 +161,7 @@ public class ParallecPojoClassTest {
         pojoClasses.add(TcpUdpRequestCreateException.class);
         pojoClasses.add(TcpUdpSshPingResourceStore.class);
         
-        
+        pojoClasses.add(ResponseHeaderMeta.class);
         
         pojoClasses.add(HttpClientStore.class);